requirements.txt
absl-py==1.0.0
argon2-cffi==21.1.0
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
cached-property==1.5.2
cachetools==5.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
classproperties==0.2.0
cloudpickle==2.0.0
cycler==0.11.0
datafiles==1.1.1
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.4
dm-tree==0.1.7
entrypoints==0.3
flatbuffers==2.0
fonttools==4.33.3
gast==0.5.3
gitman==3.0.1
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.1
graphsurgeon==0.4.5
grpcio==1.46.1
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
imutils==0.5.4
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
joblib==1.1.0
jsonschema==4.1.2
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
Markdown==3.3.7
matplotlib==3.5.2
matplotlib-inline==0.1.3
minilog==2.0.1
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.22.3
oauthlib==3.2.0
onnx-graphsurgeon==0.3.10
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.0
pandas==1.4.2
pandocfilters==1.5.0
parse==1.19.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pkg_resources==0.0.0
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.20.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2022.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.16.1
tensorflow-datasets==4.5.2
tensorflow-gpu==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
tensorflow-metadata==1.8.0
tensorflow-probability==0.16.0
tensorrt==8.0.1.6
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tomlkit==0.7.2
tornado==6.1
tqdm==4.64.0
traitlets==5.1.0
typeguard==2.13.3
typing-extensions==3.10.0.2
uff==0.6.9
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.5.1
wrapt==1.14.1
zipp==3.8.0
